+++ lib/src/bookings/data/model/booking/others_booking.dart
@@ -21,7 +21,7 @@ class OthersBooking {
   String? recommendations;
 
   factory OthersBooking.fromJson(Map<String, dynamic> json) => OthersBooking(
-        amount: json['amount'] ?? 0.0,
+        amount: json['amount'].toDouble() ?? 0.0,
         others: List<OtherServ>.from(
             json['others'].map((x) => OtherServ.fromJson(x))),
         recommendations: json['recommendations'] ?? '',
